backend_py/worker_mock/pyproject.toml
[project]
name = "worker-mock"
version = "0.0.1"
requires-python = "^3.13"
dynamic = ["dependencies"]

[tool.poetry]
requires-poetry = ">=2.1"
packages = [{ include = "worker_mock", from = "src" }]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
aiohttp = "^3.13.2"
azure-servicebus = "^7.14.3"
azure-monitor-opentelemetry = "^1.8.1"
watchfiles = "^1.1.0"
webviz-core-utils = { path = "../libs/core_utils", develop = true }
webviz-server-schemas = { path = "../libs/server_schemas", develop = true }
webviz-services = { path = "../libs/services", develop = true }
